# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Matador API
  description: Matador on-demand car rental API. Customer endpoints use JWT bearer
    tokens; admin endpoints use session cookies.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/webhooks/stripe:
    post:
      tags:
      - Webhooks
      summary: Stripe webhook
      description: Handle Stripe payment events.
      operationId: handle
      parameters:
      - name: Stripe-Signature
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: OK
...
  /api/customer/me/trips/{id}/inspections/{phase}:
    post:
      tags:
      - Customer-Trips
      summary: Submit inspection
      description: Submit an inspection with all photo URLs.
      operationId: submit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: phase
        in: path
        required: true
        schema:
          type: string
          enum:
          - PICKUP
          - DROPOFF
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitInspectionRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InspectionResponse'
  /api/customer/me/trips/{id}/inspections/{phase}/upload-url:
    post:
      tags:
      - Customer-Trips
      summary: Presign upload
      description: Get a presigned upload URL for a photo angle.
      operationId: uploadUrl
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: phase
        in: path
        required: true
        schema:
          type: string
          enum:
          - PICKUP
          - DROPOFF
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
...
  /api/customer/me/payment-methods:
    get:
      tags:
      - Customer-Profile
      summary: List payment methods
      description: Saved cards for the customer.
      operationId: list
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethodResponse'
    post:
      tags:
      - Customer-Profile
      summary: Attach card
      description: Attach a confirmed payment method to the customer.
      operationId: attach
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachPaymentMethodRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaymentMethodResponse'
  /api/customer/me/payment-methods/{id}/default:
    post:
      tags:
      - Customer-Profile
      summary: Set default card
      description: Mark a payment method as default.
      operationId: setDefault
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
...
  /api/admin/ledger/adjustments:
    post:
      tags:
      - Admin-Ledger
      summary: Manual adjustment
      description: Record a manual ledger adjustment.
      operationId: adjust
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustmentRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
                  format: uuid
  /api/admin/inspections/{id}/review:
    post:
      tags:
      - Admin-Trips
      summary: Review inspection
      description: Mark an inspection passed or flagged.
      operationId: review
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InspectionResponse'
...
  /api/admin/customers/{id}:
    get:
      tags:
      - Admin-Customers
      summary: Customer detail
      description: Full customer profile.
      operationId: detail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdminCustomerDetail'
    patch:
      tags:
      - Admin-Customers
      summary: Update status
      description: Suspend, deactivate, or reactivate a customer.
      operationId: updateStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerStatusRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdminCustomerDetail'
...
  /api/customer/me/trips:
    get:
      tags:
      - Customer-Trips
      summary: Trip history
      description: Paginated trip history.
      operationId: history
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageTripResponse'
...
  /api/admin/trips/{id}:
    get:
      tags:
      - Admin-Trips
      summary: Trip detail
      description: A single trip.
      operationId: get_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TripResponse'
  /api/admin/ledger/trips/{tripId}:
    get:
      tags:
      - Admin-Ledger
      summary: Trip ledger
      description: All ledger entries for a trip.
      operationId: byTrip
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerEntryView'
  /api/admin/ledger/customers/{customerId}:
    get:
      tags:
      - Admin-Ledger
      summary: Customer ledger
      description: Paginated ledger history for a customer.
      operationId: byCustomer
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageLedgerEntryView'
  /api/admin/inspections/{id}:
    get:
      tags:
      - Admin-Trips
      summary: Inspection detail
      description: View an inspection with photos.
      operationId: get_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InspectionResponse'
...
  /api/customer/me/payment-methods/{id}:
    delete:
      tags:
      - Customer-Profile
      summary: Detach card
      description: Remove a saved payment method.
      operationId: detach
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
...
components:
  schemas:
...
    PhotoSubmission:
      required:
      - angle
      - url
      type: object
      properties:
        angle:
          type: string
          enum:
          - FRONT
          - BACK
          - LEFT
          - RIGHT
          - INTERIOR_FRONT
          - INTERIOR_REAR
          - ODOMETER
        url:
          type: string
        capturedAt:
          type: string
          format: date-time
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        fileSizeBytes:
          type: integer
          format: int32
        widthPx:
          type: integer
          format: int32
        heightPx:
          type: integer
          format: int32
    SubmitInspectionRequest:
      required:
      - photos
      type: object
      properties:
        odometerMiles:
          type: integer
          format: int32
        fuelChargePercent:
          type: integer
          format: int32
        notes:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PhotoSubmission'
    InspectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        phase:
          type: string
          enum:
          - PICKUP
          - DROPOFF
        odometerMiles:
          type: integer
          format: int32
        fuelChargePercent:
          type: integer
          format: int32
        reviewStatus:
          type: string
          enum:
          - PASSED
          - FLAGGED
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PhotoView'
    PhotoView:
      type: object
      properties:
        angle:
          type: string
          enum:
          - FRONT
          - BACK
          - LEFT
          - RIGHT
          - INTERIOR_FRONT
          - INTERIOR_REAR
          - ODOMETER
        url:
          type: string
    UploadUrlRequest:
      required:
      - angle
      type: object
      properties:
        angle:
          type: string
          enum:
          - FRONT
          - BACK
          - LEFT
          - RIGHT
          - INTERIOR_FRONT
          - INTERIOR_REAR
          - ODOMETER
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        publicUrl:
          type: string
        key:
          type: string
...
    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - IN_PROGRESS
          - ENDED_PENDING_INSPECTION
          - CLOSED
        actualPickupAt:
          type: string
          format: date-time
        actualDropoffAt:
          type: string
          format: date-time
        odometerStart:
          type: integer
          format: int32
        odometerEnd:
          type: integer
          format: int32
        milesDriven:
          type: integer
          format: int32
        finalChargesCents:
          type: integer
          format: int64
    AttachPaymentMethodRequest:
      required:
      - stripePaymentMethodId
      type: object
      properties:
        stripePaymentMethodId:
          type: string
        isDefault:
          type: boolean
    PaymentMethodResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        brand:
          type: string
        last4:
          type: string
        expMonth:
          type: integer
          format: int32
        expYear:
          type: integer
          format: int32
        isDefault:
          type: boolean
...
    AdjustmentRequest:
      required:
      - customerId
      - reason
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        amountCents:
          type: integer
          format: int64
        reason:
          type: string
    ReviewRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          enum:
          - PASSED
          - FLAGGED
        notes:
          type: string
...
    UpdateCustomerStatusRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DEACTIVATED
        reason:
          type: string
    AdminCustomerDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        licenseNumber:
          type: string
        licenseState:
          type: string
        licenseExpiresOn:
          type: string
          format: date
        verificationStatus:
          type: string
          enum:
          - UNVERIFIED
          - IN_PROGRESS
          - VERIFIED
          - REJECTED
          - EXPIRED
        verificationCompletedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DEACTIVATED
        stripeCustomerId:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageTripResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/TripResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
    LedgerEntryView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
        entryType:
          type: string
          enum:
          - DEPOSIT_HELD
          - DEPOSIT_CAPTURED
          - DEPOSIT_RELEASED
          - RENTAL_CHARGED
          - OVERAGE_CHARGED
          - INCIDENT_CHARGED
          - REFUND_ISSUED
          - DISPUTE_CHARGEBACK
          - MANUAL_ADJUSTMENT
        amountCents:
          type: integer
          format: int64
        currency:
          type: string
        description:
          type: string
    PageLedgerEntryView:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntryView'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: Customer JWT access token.
      scheme: bearer
      bearerFormat: JWT
    session-cookie:
      type: apiKey
      description: Admin session cookie.
      name: SESSION
      in: cookie